Refresh stale generated snapshots

+++ tests/byte_char_comparisons/main.rs
@@ -54,8 +54,8 @@ pub fn classify_op(op: Rc<RefCell<Option<u8>>>) -> Rc<RefCell<Option<String>>> {
 }
 
 fn main() {
-    println!("{} {}", format!("{}", "digit 5:".to_string()), format!("{}", is_digit(Rc::new(RefCell::new(Some(('5' as u8)))))));
-    println!("{} {}", format!("{}", "digit x:".to_string()), format!("{}", is_digit(Rc::new(RefCell::new(Some(('x' as u8)))))));
+    println!("{} {}", format!("{}", "digit 5:".to_string()), format!("{}", is_digit(Rc::new(RefCell::new(Some(('5' as i32) as u8))))));
+    println!("{} {}", format!("{}", "digit x:".to_string()), format!("{}", is_digit(Rc::new(RefCell::new(Some(('x' as i32) as u8))))));
     println!("{} {}", format!("{}", "version v1:".to_string()), format!("{}", starts_with_v(Rc::new(RefCell::new(Some("v1.0.0".to_string()))))));
     println!("{} {}", format!("{}", "version x1:".to_string()), format!("{}", starts_with_v(Rc::new(RefCell::new(Some("x1.0.0".to_string()))))));
     let mut path = Rc::new(RefCell::new(Some(Vec::<u8>::new())));
