+++ tests/cross_file_named_uint_const_peers/main.rs
@@ -14,7 +14,7 @@ pub fn decode(ver: Rc<RefCell<Option<u32>>>, flags: Rc<RefCell<Option<u32>>>) ->
 }
 
 fn main() {
-    println!("{}", format!("{}", decode(Rc::new(RefCell::new(Some(1))), Rc::new(RefCell::new(Some(1))))));
-    println!("{}", format!("{}", decode(Rc::new(RefCell::new(Some(3))), Rc::new(RefCell::new(Some(1))))));
-    println!("{}", format!("{}", decode(Rc::new(RefCell::new(Some(1))), Rc::new(RefCell::new(Some(0))))));
+    println!("{}", format!("{}", decode(Rc::new(RefCell::new(Some(1 as u32))), Rc::new(RefCell::new(Some(1 as u32))))));
+    println!("{}", format!("{}", decode(Rc::new(RefCell::new(Some(3 as u32))), Rc::new(RefCell::new(Some(1 as u32))))));
+    println!("{}", format!("{}", decode(Rc::new(RefCell::new(Some(1 as u32))), Rc::new(RefCell::new(Some(0 as u32))))));
 }
